ecommerce_project/services/order/requirements.txt
fastapi
uvicorn
python-dotenv
pika
aio_pika
psycopg2-binary